@@ -10,8 +10,8 @@ class Preview(BaseModel):
 
 class Transcription(BaseModel):
     status: str
-    locale: Optional[str]
-    preview: Optional[Preview]
+    locale: Optional[str] = None
+    preview: Optional[Preview] = None
 
 
 class FileObject(BaseModel):
@@ -37,42 +37,43 @@ class FileObject(BaseModel):
     url_private: str
     url_private_download: str
     media_display_type: str
-    original_w: Optional[int]
-    original_h: Optional[int]
+    original_w: Optional[int] = None
+    original_h: Optional[int] = None
+
     # thumb
-    thumb_64: Optional[str]
-    thumb_80: Optional[str]
-    thumb_360: Optional[str]
-    thumb_360_w: Optional[int]
-    thumb_360_h: Optional[int]
-    thumb_480: Optional[str]
-    thumb_480_w: Optional[int]
-    thumb_480_h: Optional[int]
-    thumb_160: Optional[str]
-    thumb_720: Optional[str]
-    thumb_720_w: Optional[int]
-    thumb_720_h: Optional[int]
-    thumb_800: Optional[str]
-    thumb_800_w: Optional[int]
-    thumb_800_h: Optional[int]
-    thumb_960: Optional[str]
-    thumb_960_w: Optional[int]
-    thumb_960_h: Optional[int]
-    thumb_1024: Optional[str]
-    thumb_1024_w: Optional[int]
-    thumb_1024_h: Optional[int]
-    thumb_tiny: Optional[str]
-    thumb_pdf: Optional[str]
-    thumb_pdf_w: Optional[int]
-    thumb_pdf_h: Optional[int]
+    thumb_64: Optional[str] = None
+    thumb_80: Optional[str] = None
+    thumb_360: Optional[str] = None
+    thumb_360_w: Optional[int] = None
+    thumb_360_h: Optional[int] = None
+    thumb_480: Optional[str] = None
+    thumb_480_w: Optional[int] = None
+    thumb_480_h: Optional[int] = None
+    thumb_160: Optional[str] = None
+    thumb_720: Optional[str] = None
+    thumb_720_w: Optional[int] = None
+    thumb_720_h: Optional[int] = None
+    thumb_800: Optional[str] = None
+    thumb_800_w: Optional[int] = None
+    thumb_800_h: Optional[int] = None
+    thumb_960: Optional[str] = None
+    thumb_960_w: Optional[int] = None
+    thumb_960_h: Optional[int] = None
+    thumb_1024: Optional[str] = None
+    thumb_1024_w: Optional[int] = None
+    thumb_1024_h: Optional[int] = None
+    thumb_tiny: Optional[str] = None
+    thumb_pdf: Optional[str] = None
+    thumb_pdf_w: Optional[int] = None
+    thumb_pdf_h: Optional[int] = None
     # thumb end
 
     # audio
-    subtype: Optional[str]
-    duration_ms: Optional[int]
-    aac: Optional[str]
-    audio_wave_samples: Optional[List[int]]
-    transcription: Optional[Transcription]
+    subtype: Optional[str] = None
+    duration_ms: Optional[int] = None
+    aac: Optional[str] = None
+    audio_wave_samples: Optional[List[int]] = None
+    transcription: Optional[Transcription] = None
     # audio end
 
     permalink: str
